refactor: extract profit loss report components for better maintainability
- Split ProfitLoss page into separate components:
  - SummaryCards: Shows profit/loss overview
  - PerformanceSummary: Displays performance metrics
  - ProductBreakdownColumns: Product breakdown configuration
- Reduce main page component complexity
